fix(Auth): Updated to v5 next-auth & preliminary fixes
