+++ guida/package.json
@@ -17,7 +17,7 @@
     "bcryptjs": "^2.4.3",
     "guida": "file:",
     "next": "14.2.13",
-    "next-auth": "^4.24.8",
+    "next-auth": "^5.0.0-beta.22",
     "react": "^18",
     "react-dom": "^18"
   },
