+++ guida/src/app/api/posts/route.ts
@@ -22,6 +22,7 @@ interface PostRequestPost {
  * @constructor
  */
 export async function GET(request : Request) {
+    console.log('RAN THAT SHIT')
     try {
         const { searchParams } = new URL(request.url)
         const postId = searchParams.get('id')
