+++ guida/src/routes.ts
@@ -1,0 +1,30 @@
+/**
+ * An array of routes that are accessible to the public
+ * These routes do not require authentication
+ * @type {string[]}
+ */
+export const publicRoutes: string[] = [
+
+];
+
+/**
+ * An array of routes that are used for authentication
+ * These routes will redirect logged in users to DEFAULT_LOGIN_REDIRECT
+ * @type {string[]}
+ */
+export const authRoutes: string[] = [
+    "/login"
+]
+
+/**
+ * The prefix for API authentication routes
+ * Routes that start with this prefix are used for API authentication purposes
+ * @type {string}
+ */
+export const apiAuthPrefix: string = '/api/auth'
+
+/**
+ * The default redirect path after logging in
+ * @type {string}
+ */
+export const DEFAULT_LOGIN_REDIRECT: string = '/'
